+++ src/components/DialogUI/DialogUI.js
@@ -52,18 +52,24 @@ function DialogUI({
       open={open}
       fullWidth={true}
       maxWidth={maxWidth}
-      onClose={onClose}
+      onClose={(e) => {
+        e.stopPropagation();
+        e.preventDefault();
+        onClose(e);
+      }}
       PaperComponent={PaperComponent}
       aria-labelledby="draggable-dialog-title"
     >
       {title && (
-        <DialogTitle
-          className={
-            draggable ? classesLocalTitle.draggable : classesLocalTitle.undraggable
-          }
-          id={draggable ? 'draggable-dialog-title' : ''}
-        >
-          {title}
+        <>
+          <DialogTitle
+            className={
+              draggable ? classesLocalTitle.draggable : classesLocalTitle.undraggable
+            }
+            id={draggable ? 'draggable-dialog-title' : ''}
+          >
+            {title}
+          </DialogTitle>
           {isClosable && (
             <IconButton
               aria-label="close"
@@ -73,7 +79,7 @@ function DialogUI({
               <CloseIcon />
             </IconButton>
           )}
-        </DialogTitle>
+        </>
       )}
       <DialogContent className={classes.content}>{children}</DialogContent>
       {(primary?.onClick || secondary?.onClick) && (